+++ arctic/generics.py
@@ -1,5 +1,7 @@
 from __future__ import (division, unicode_literals)
 from collections import OrderedDict
+import calendar
+import json
 
 import extra_views
 from django.conf import settings
@@ -10,6 +12,7 @@ from django.core.exceptions import (FieldDoesNotExist, ImproperlyConfigured,
                                     ValidationError)
 from django.core.paginator import InvalidPage
 from django.urls import (NoReverseMatch, reverse)
+from django.utils.html import mark_safe
 from django.db.models.deletion import (Collector, ProtectedError)
 from django.forms.widgets import Media
 from django.http import Http404
@@ -189,6 +192,19 @@ class View(RoleAuthentication, base.View):
                                        get_language())[0]
         dtformats['SHORT_DATETIME'] = get_format('DATETIME_INPUT_FORMATS',
                                                  get_language())[0]
+        dtformats['DAYS'] = mark_safe(json.dumps(
+            [_(day) for day in calendar.day_name]))
+        dtformats['DAYS_SHORT'] = mark_safe(json.dumps(
+            [_(day) for day in calendar.day_abbr]))
+        dtformats['DAYS_MIN'] = mark_safe(json.dumps(
+            [_(day)[:2] for day in calendar.day_abbr]))
+        dtformats['MONTHS'] = mark_safe(json.dumps(
+            [_(day).capitalize() for day in calendar.month_name[1:]]))
+        dtformats['MONTHS_SHORT'] = mark_safe(json.dumps(
+            [_(day) for day in calendar.month_abbr[1:]]))
+        dtformats['TODAY'] = _('Today')
+        dtformats['CLEAR'] = _('Clear')
+        dtformats['FIRST_DAY'] = calendar.firstweekday()
 
         return dtformats
 
@@ -407,7 +423,7 @@ class ListView(View, ListMixin, base.ListView):
 
     def get_list_items(self, objects):  # noqa: C901
         self.has_action_links = False
-        has_confirm_links = hasattr(self, 'confirm_links')
+        has_modal_links = hasattr(self, 'modal_links')
         items = []
         if not self.get_fields():
             for obj in objects:
@@ -442,8 +458,10 @@ class ListView(View, ListMixin, base.ListView):
                 if field_name in field_links.keys():
                     field['url'] = self._reverse_field_link(
                         field_links[field_name], obj)
-                    self.add_confirm_link(has_confirm_links, field,
+                    self.add_confirm_link(has_modal_links, field,
                                           field_links[field_name])
+                    field['confirm'] = self.get_confirm_link(
+                        field_links[field_name], obj)
                 if field_name in field_classes:
                     field['class'] = field_classes[field_name]
                 row.append(field)
@@ -460,8 +478,8 @@ class ListView(View, ListMixin, base.ListView):
         return items
 
     def add_confirm_link(self, has_confirm_link, field, field_url_name):
-        if has_confirm_link and field_url_name in self.confirm_links:
-            field['confirm'] = self.confirm_links[field_url_name]
+        if has_confirm_link and field_url_name in self.modal_links:
+            field['confirm'] = self.modal_links[field_url_name]
 
     def get_field_value(self, field_name, obj):
         # first try to find a virtual field
@@ -516,6 +534,8 @@ class ListView(View, ListMixin, base.ListView):
         context['has_action_links'] = self.has_action_links
         context['sorting_url'] = self.sorting_url
         context['tool_links_icon'] = self.get_tool_links_icon()
+        context['tool_links_collapse'] = self.tool_links_collapse - \
+            int(len(context['tool_links']) == self.tool_links_collapse + 1)
         context['simple_search_form'] = self.simple_search_form
         context['advanced_search_form'] = self.advanced_search_form
         return context
@@ -563,12 +583,10 @@ class DataListView(TemplateView, ListMixin):
         # self.has_action_links is set in get_list_items
         context['has_action_links'] = self.has_action_links
         context['tool_links_icon'] = self.get_tool_links_icon()
-        if self.get_simple_search_form_class():
-            context['simple_search_form'] = \
-                self.get_simple_search_form_class()(data=self.request.GET)
-        if self.get_advanced_search_form_class():
-            context['advanced_search_form'] = \
-                self.get_advanced_search_form_class()(data=self.request.GET)
+        context['tool_links_collapse'] = self.tool_links_collapse - \
+            int(len(context['tool_links']) == self.tool_links_collapse + 1)
+        context['simple_search_form'] = self.simple_search_form
+        context['advanced_search_form'] = self.advanced_search_form
 
         return context
 
@@ -702,7 +720,7 @@ class DataListView(TemplateView, ListMixin):
 class CreateView(FormMediaMixin, View, SuccessMessageMixin,
                  FormMixin, extra_views.CreateWithInlinesView):
     template_name = 'arctic/base_create_update.html'
-    success_message = _('%(object)s was created successfully')
+    success_message = _('"%(object)s" was successfully created')
 
     def get_page_title(self):
         if not self.page_title:
@@ -711,15 +729,16 @@ class CreateView(FormMediaMixin, View, SuccessMessageMixin,
 
     def get_context_data(self, **kwargs):
         context = super(CreateView, self).get_context_data(**kwargs)
-        context['links'] = self.get_links()
+        context['actions'] = self.get_actions()
         context['layout'] = self.get_layout()
+        context['cancel_url'] = self.get_cancel_url()
         return context
 
 
 class UpdateView(FormMediaMixin, SuccessMessageMixin, FormMixin, View,
                  LinksMixin, extra_views.UpdateWithInlinesView):
     template_name = 'arctic/base_create_update.html'
-    success_message = _('%(object)s was updated successfully')
+    success_message = _('"%(object)s" was successfully updated')
 
     links = None             # Optional links such as list of linked items
     readonly_fields = None   # Optional list of readonly fields
@@ -731,8 +750,9 @@ class UpdateView(FormMediaMixin, SuccessMessageMixin, FormMixin, View,
 
     def get_context_data(self, **kwargs):
         context = super(UpdateView, self).get_context_data(**kwargs)
-        context['links'] = self.get_links()
+        context['actions'] = self.get_actions()
         context['layout'] = self.get_layout()
+        context['cancel_url'] = self.get_cancel_url()
         return context
 
 
@@ -742,17 +762,27 @@ class FormView(FormMediaMixin, View, SuccessMessageMixin, FormMixin,
 
     def get_context_data(self, **kwargs):
         context = super(FormView, self).get_context_data(**kwargs)
-        context['links'] = self.get_links()
+        context['actions'] = self.get_actions()
         context['layout'] = self.get_layout()
+        context['cancel_url'] = self.get_cancel_url()
         return context
 
 
 class DeleteView(View, base.DeleteView):
     template_name = 'arctic/base_confirm_delete.html'
-    redirect = False
+    redirect = True
+
+    @staticmethod
+    def confirm_dialog():
+        return {
+            'title': _('Delete "{obj}"'),
+            'message': _('Are you sure you want to delete this?'),
+            'ok': _('Delete'),
+            'cancel': _('Cancel'),
+        }
 
     def get_success_message(self, obj):
-        return _('"{}" has been successfully deleted.').format(str(obj))
+        return _('"{}" was successfully deleted.').format(str(obj))
 
     def get(self, request, *args, **kwargs):
         """
@@ -774,7 +804,8 @@ class DeleteView(View, base.DeleteView):
         if can_delete and self.redirect:
             messages.success(request, self.get_success_message(self.object))
             self.delete(request, *args, **kwargs)
-            return redirect(self.success_url)
+            return redirect(self.request.META.get('HTTP_REFERER',
+                                                  self.get_success_url()))
 
         context = self.get_context_data(object=self.object,
                                         can_delete=can_delete,