+++ arctic/models.py
@@ -1,5 +1,6 @@
 from django.conf import settings
 from django.db import models
+from django.utils.translation import ugettext_lazy as _
 
 
 class UserRole(models.Model):
@@ -10,10 +11,11 @@ class UserRole(models.Model):
 
     class Meta:
         swappable = 'ARCTIC_USER_ROLE_MODEL'
+        ordering = ['user']
 
 
 class Role(models.Model):
-    name = models.CharField('Role', max_length=100, unique=True)
+    name = models.CharField(_('Role'), max_length=100, unique=True)
     is_active = models.BooleanField(default=True)
 
     def __str__(self):