+++ tests/test_forms.py
@@ -1,28 +1,10 @@
 from django.db.models import Q
-from django.http import HttpRequest
-from django.template import Context, Template
-from bs4 import BeautifulSoup
+# from django.http import HttpRequest
+# from django.template import Context, Template
+# from bs4 import BeautifulSoup
 
-from arctic.forms import SimpleSearchForm, QuickFiltersFormMixin
+from arctic.forms import SimpleSearchForm
 from arctic.widgets import QuickFiltersSelect
-
-
-class FiltersForm(QuickFiltersFormMixin, SimpleSearchForm):
-    FILTER_BUTTONS = (
-        ('published', 'Is published'),
-        ('find_rabbit', 'Rabbit')
-    )
-    filters_query_name = 'my_filters'
-
-    def get_quick_filter_query(self):
-        value = self.cleaned_data.get(self.filters_query_name)
-
-        if value == 'published':
-            return Q(published=True)
-        elif value == 'rabbit':
-            return Q(description__icontains='rabbit')
-        else:
-            return Q()
 
 
 def test_simple_search_form():
@@ -48,45 +30,47 @@ def test_filters_widget_attr():
     assert widget.template_name == 'arctic/widgets/quick_filters_select.html'
 
 
-def test_filters_form_field():
-    form = FiltersForm()
+# def test_form_rendering_with_request_get_args():
+#     request = HttpRequest()
+#     request.GET['search'] = 'cats'
+#     request.GET['my_filters'] = 'published'
 
-    assert form.fields.get('my_filters')
+#     form = FiltersForm(data=request.GET)
+#     assert form.is_valid()
 
-    widget = form.fields['my_filters'].widget
-    assert isinstance(widget, QuickFiltersSelect)
+#     template = Template('{{ form }}').render(Context({"form": form}))
 
-    assert form.get_quick_filters_field().field == form.fields['my_filters']
+#     soup = BeautifulSoup(template, 'html.parser')
 
-    assert not form.fields['my_filters'].required
-    assert not form.fields['my_filters'].choices == form.FILTER_BUTTONS
+#     filters_inputs = soup.find_all('input', {'type': 'checkbox'})
+#     filters_buttons = soup.find_all('label', {'class': 'btn'})
+
+#     assert len(filters_inputs) == len(form.FILTER_BUTTONS)
+#     assert len(filters_buttons) == len(form.FILTER_BUTTONS)
+
+#     for inp, btn, choice in zip(
+#             filters_inputs,
+#             filters_buttons,
+#             form.FILTER_BUTTONS):
+#         assert btn.text.strip() == choice[1]
+#         assert inp.get('value') == choice[0]
+#         assert inp.get('name') == 'my_filters'
+
+#     # check if we marked selected filter
+#     assert filters_inputs[0].attrs.get('checked') == ''
 
 
-def test_form_rendering_with_request_get_args():
-    request = HttpRequest()
-    request.GET['search'] = 'cats'
-    request.GET['my_filters'] = 'published'
+# def test_filters_form_rendering_select_one():
+#     request = HttpRequest()
+#     request.GET['search'] = 'cats'
+#     request.GET['my_filters'] = 'published'
 
-    form = FiltersForm(data=request.GET)
-    assert form.is_valid()
+#     form = FiltersFormSelectOne(data=request.GET)
+#     assert form.is_valid()
 
-    template = Template('{{ form }}').render(Context({"form": form}))
+#     template = Template('{{ form }}').render(Context({"form": form}))
 
-    soup = BeautifulSoup(template, 'html.parser')
+#     soup = BeautifulSoup(template, 'html.parser')
+#     filters_inputs = soup.find_all('input', {'type': 'checkbox'})
 
-    filters_inputs = soup.find_all('input', {'type': 'checkbox'})
-    filters_buttons = soup.find_all('label', {'class': 'btn'})
-
-    assert len(filters_inputs) == len(form.FILTER_BUTTONS)
-    assert len(filters_buttons) == len(form.FILTER_BUTTONS)
-
-    for inp, btn, choice in zip(
-            filters_inputs,
-            filters_buttons,
-            form.FILTER_BUTTONS):
-        assert btn.text.strip() == choice[1]
-        assert inp.get('value') == choice[0]
-        assert inp.get('name') == 'my_filters'
-
-    # check if we marked selected filter
-    assert filters_inputs[0].attrs.get('checked') == ''
+#     assert filters_inputs[0].attrs.get('checked') is None